requirements.txt
openai
arxiv
faiss-cpu
sentence-transformers
requests
PyMuPDF  
streamlit
